+++ src/Common.Media/Drawing/Models/DTO/TextImageInputDto.cs
@@ -18,14 +18,14 @@ public class TextImageInputDto
     public TextOptions? Options { get; set; }
 
     public LengthUnit? DimensionUnit { get; set; }
-    public double? Width { get; set; }
-    public double? Height { get; set; }
-    public double? Margin { get; set; }
+    public float? Width { get; set; }
+    public float? Height { get; set; }
+    public float? Margin { get; set; }
     public ImagePosition? PositionType { get; set; }
-    public double? Left { get; set; }
-    public double? Top { get; set; }
-    public double? Right { get; set; }
-    public double? Bottom { get; set; }
-    public double? Rotation { get; set; }
-    public double? Opacity { get; set; }
+    public float? Left { get; set; }
+    public float? Top { get; set; }
+    public float? Right { get; set; }
+    public float? Bottom { get; set; }
+    public float? Rotation { get; set; }
+    public float? Opacity { get; set; }
 }
